Changed 'accum' in GameLoop.java from int to float to correctly report the update alpha in the Game.paint() method.

+++ android/src/playn/android/GameLoop.java
@@ -29,7 +29,7 @@ public class GameLoop implements Runnable {
   private long timeOffset = System.currentTimeMillis();
 
   private int updateRate;
-  private int accum;
+  private float accum;
   private int lastTime;
 
   private float totalTime;
